+++ app/controllers/admin/auth_controller.rb
@@ -17,7 +17,8 @@ module Admin
         session[:user_id] = admin_session.user_id
         session[:organisation_id] = admin_session.organisation_id
         session[:role_codes] = admin_session.role_codes
-        redirect_to admin_root_path
+
+        redirect_to session.delete(:requested_admin_path) || admin_root_path
       else
         render "failure", status: :unauthorized
       end