+++ spec/features/admin_sessions_spec.rb
@@ -1,16 +1,28 @@
 require "rails_helper"
 
-RSpec.feature "Admin sessions" do
-  before do
+RSpec.feature "Admin session management" do
+  scenario "A user is redirected to the admin root path after sign in" do
     sign_in_to_admin_with_role(AdminSession::SERVICE_OPERATOR_DFE_SIGN_IN_ROLE_CODE)
-  end
 
-  scenario "Redirected to admin page after signing in" do
     expect(page).to have_link("Sign out")
+    expect(current_path).to eql(admin_root_path)
   end
 
-  scenario "Signing out" do
+  scenario "A signed in user can sign out" do
+    sign_in_to_admin_with_role(AdminSession::SERVICE_OPERATOR_DFE_SIGN_IN_ROLE_CODE)
+
     click_on "Sign out"
     expect(page).to have_content("You've been signed out")
+    expect(current_path).to eql(admin_sign_in_path)
+  end
+
+  scenario "A user is redirected to their original url after sign in" do
+    visit admin_claims_path
+
+    expect(current_path).to eql(admin_sign_in_path)
+
+    sign_in_to_admin_with_role(AdminSession::SERVICE_OPERATOR_DFE_SIGN_IN_ROLE_CODE)
+
+    expect(current_path).to eql(admin_claims_path)
   end
 end